backend/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.6
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
